Calendar Implemented:
Users are now able to see the tasks they have created displayed on the calendar. It took a while and I keept a comment of the error I was getting.

Summarised, the issue was that the calendarTasks was being pouplated with all the tasks and thats why the dots show on the dates. However, when a date was selected, the calendarTasks would clear and only be populated with the task(if any) on that date which caused the dot to dissappear.

My solution, I created calendarTasksView, the purpose of this is so this list will be used to display the tasks for the selected day. This means that calendarTasks remains untouched since it being populated which means the dots will always be displayed and the tasks change successfully.