Notes Pinned on Welcome Page:
Before, there was a ViewModel which I was not familiar on how to use. It works using LiveData instead of MutableList or List. Using this would result in changing a lot more aspects of my code so I opted not to as the data is always refreshed.

Pinned notes show in the welcome screen and now it shows a Welcome message as well.